+++ app/src/main/java/com/fibelatti/photowidget/home/HomeScreen.kt
@@ -72,6 +72,7 @@ fun HomeScreen(
     val existingWidgetMenuSheetState = rememberAppSheetState()
     val removedWidgetSheetState = rememberAppSheetState()
     val invalidWidgetSheetState = rememberAppSheetState()
+    val draftWidgetSheetState = rememberAppSheetState()
     val appAppearanceSheetState = rememberAppSheetState()
     val appColorsSheetState = rememberAppSheetState()
 
@@ -121,6 +122,11 @@ fun HomeScreen(
                 data = appWidgetId,
             )
         },
+        onDraftWidgetClick = { appWidgetId ->
+            draftWidgetSheetState.showBottomSheet(
+                data = appWidgetId,
+            )
+        },
         onDefaultsClick = {
             localContext.startActivity(Intent(localContext, WidgetDefaultsActivity::class.java))
         },
@@ -166,6 +172,11 @@ fun HomeScreen(
         onDelete = homeViewModel::deleteWidget,
     )
 
+    DraftWidgetBottomSheet(
+        sheetState = draftWidgetSheetState,
+        onDelete = homeViewModel::deleteWidget,
+    )
+
     AppAppearanceBottomSheet(
         sheetState = appAppearanceSheetState,
     )
@@ -183,6 +194,7 @@ fun HomeScreen(
     onCurrentWidgetClick: (appWidgetId: Int, canSync: Boolean, canLock: Boolean, isLocked: Boolean) -> Unit,
     onRemovedWidgetClick: (appWidgetId: Int, PhotoWidgetStatus) -> Unit,
     onInvalidWidgetClick: (appWidgetId: Int) -> Unit,
+    onDraftWidgetClick: (appWidgetId: Int) -> Unit,
     onDefaultsClick: () -> Unit,
     onAppearanceClick: () -> Unit,
     onColorsClick: () -> Unit,
@@ -245,6 +257,7 @@ fun HomeScreen(
                         onCurrentWidgetClick = onCurrentWidgetClick,
                         onRemovedWidgetClick = onRemovedWidgetClick,
                         onInvalidWidgetClick = onInvalidWidgetClick,
+                        onDraftWidgetClick = onDraftWidgetClick,
                     )
                 }
 
@@ -364,6 +377,7 @@ private fun HomeScreenPreview() {
             onCurrentWidgetClick = { _, _, _, _ -> },
             onRemovedWidgetClick = { _, _ -> },
             onInvalidWidgetClick = {},
+            onDraftWidgetClick = {},
             onDefaultsClick = {},
             onAppearanceClick = {},
             onColorsClick = {},